refactor: 의존성 주입 방법 생성자 방법으로 변경
