+++ src/main/java/woowacourse/auth/domain/Customer.java
@@ -1,6 +1,6 @@
 package woowacourse.auth.domain;
 
-import static lombok.EqualsAndHashCode.*;
+import static lombok.EqualsAndHashCode.Include;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -11,19 +11,19 @@ import lombok.Getter;
 @Getter
 public class Customer {
 
-	@Include
-	private Long id;
-	private final String email;
-	private final String password;
-	private final String nickname;
+    @Include
+    private Long id;
+    private final String email;
+    private final String password;
+    private final String nickname;
 
-	public Customer(String email, String password, String nickname) {
-		this.email = email;
-		this.password = password;
-		this.nickname = nickname;
-	}
+    public Customer(String email, String password, String nickname) {
+        this.email = email;
+        this.password = password;
+        this.nickname = nickname;
+    }
 
-	public boolean isSamePassword(String password) {
-		return this.password.equals(password);
-	}
+    public boolean isSamePassword(String password) {
+        return this.password.equals(password);
+    }
 }
