+++ src/main/java/woowacourse/auth/dto/token/TokenRequest.java
@@ -1,9 +1,8 @@
-package woowacourse.auth.dto;
+package woowacourse.auth.dto.token;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -16,6 +15,6 @@ public class TokenRequest {
     private String email;
     @NotBlank
     @Pattern(regexp = "(?=.*[0-9])(?=.*[a-zA-Z])(?=.*\\W)(?=\\S+$).{4,20}",
-        message = "비밀번호 형식이 올바르지 않습니다.")
+            message = "비밀번호 형식이 올바르지 않습니다.")
     private String password;
 }
