+++ src/main/java/woowacourse/auth/exception/InvalidAuthException.java
@@ -1,11 +1,11 @@
 package woowacourse.auth.exception;
 
 public class InvalidAuthException extends RuntimeException {
-	public InvalidAuthException() {
-		super();
-	}
+    public InvalidAuthException() {
+        super();
+    }
 
-	public InvalidAuthException(String message) {
-		super(message);
-	}
+    public InvalidAuthException(String message) {
+        super(message);
+    }
 }
