+++ src/main/java/woowacourse/auth/exception/InvalidCustomerException.java
@@ -1,11 +1,11 @@
 package woowacourse.auth.exception;
 
 public class InvalidCustomerException extends RuntimeException {
-	public InvalidCustomerException() {
-		super();
-	}
+    public InvalidCustomerException() {
+        super();
+    }
 
-	public InvalidCustomerException(String message) {
-		super(message);
-	}
+    public InvalidCustomerException(String message) {
+        super(message);
+    }
 }
