+++ src/main/java/woowacourse/auth/ui/AuthController.java
@@ -1,7 +1,6 @@
 package woowacourse.auth.ui;
 
 import javax.validation.Valid;
-
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -9,26 +8,27 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import lombok.RequiredArgsConstructor;
 import woowacourse.auth.application.AuthService;
-import woowacourse.auth.dto.TokenRequest;
-import woowacourse.auth.dto.TokenResponse;
+import woowacourse.auth.dto.token.TokenRequest;
+import woowacourse.auth.dto.token.TokenResponse;
 
 @RestController
 @RequestMapping("/auth")
-@RequiredArgsConstructor
 public class AuthController {
 
-	private final AuthService authService;
+    private final AuthService authService;
 
-	@PostMapping("/login")
-	public ResponseEntity<TokenResponse> login(@RequestBody @Valid TokenRequest tokenRequest) {
-		return ResponseEntity.ok(authService.login(tokenRequest));
-	}
+    public AuthController(AuthService authService) {
+        this.authService = authService;
+    }
 
-	@ExceptionHandler
-	public ResponseEntity<String> loginExceptionHandler(RuntimeException exception) {
-		return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(exception.getMessage());
-	}
+    @PostMapping("/login")
+    public ResponseEntity<TokenResponse> login(@RequestBody @Valid TokenRequest tokenRequest) {
+        return ResponseEntity.ok(authService.login(tokenRequest));
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<String> loginExceptionHandler(RuntimeException exception) {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(exception.getMessage());
+    }
 }
