+++ src/main/java/woowacourse/auth/ui/AuthControllerAdvice.java
@@ -1,9 +1,7 @@
 package woowacourse.auth.ui;
 
 import java.util.List;
-
 import javax.validation.ConstraintViolationException;
-
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,48 +11,47 @@ import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-
 import woowacourse.auth.exception.InvalidAuthException;
 import woowacourse.auth.exception.InvalidCustomerException;
 
 @RestControllerAdvice(basePackages = "woowacourse.auth")
 public class AuthControllerAdvice {
 
-	@ExceptionHandler
-	public ResponseEntity<String> loginExceptionHandler(InvalidAuthException exception) {
-		return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(exception.getMessage());
-	}
+    @ExceptionHandler
+    public ResponseEntity<String> loginExceptionHandler(InvalidAuthException exception) {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(exception.getMessage());
+    }
 
-	@ExceptionHandler(EmptyResultDataAccessException.class)
-	public ResponseEntity<String> handle() {
-		return ResponseEntity.badRequest().body("존재하지 않는 데이터 요청입니다.");
-	}
+    @ExceptionHandler(EmptyResultDataAccessException.class)
+    public ResponseEntity<String> handle() {
+        return ResponseEntity.badRequest().body("존재하지 않는 데이터 요청입니다.");
+    }
 
-	@ExceptionHandler({MethodArgumentNotValidException.class})
-	public ResponseEntity<String> handleInvalidRequest(final BindingResult bindingResult) {
-		final List<FieldError> fieldErrors = bindingResult.getFieldErrors();
-		final FieldError mainError = fieldErrors.get(0);
+    @ExceptionHandler({MethodArgumentNotValidException.class})
+    public ResponseEntity<String> handleInvalidRequest(final BindingResult bindingResult) {
+        final List<FieldError> fieldErrors = bindingResult.getFieldErrors();
+        final FieldError mainError = fieldErrors.get(0);
 
-		return ResponseEntity.badRequest().body(mainError.getDefaultMessage());
-	}
+        return ResponseEntity.badRequest().body(mainError.getDefaultMessage());
+    }
 
-	@ExceptionHandler({
-		HttpMessageNotReadableException.class,
-		ConstraintViolationException.class,
-	})
-	public ResponseEntity<String> handleInvalidRequest(final RuntimeException e) {
-		return ResponseEntity.badRequest().body(e.getMessage());
-	}
+    @ExceptionHandler({
+            HttpMessageNotReadableException.class,
+            ConstraintViolationException.class,
+    })
+    public ResponseEntity<String> handleInvalidRequest(final RuntimeException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
 
-	@ExceptionHandler({
-		InvalidCustomerException.class,
-	})
-	public ResponseEntity<String> handleInvalidAccess(final RuntimeException e) {
-		return ResponseEntity.badRequest().body(e.getMessage());
-	}
+    @ExceptionHandler({
+            InvalidCustomerException.class,
+    })
+    public ResponseEntity<String> handleInvalidAccess(final RuntimeException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
 
-	@ExceptionHandler
-	public ResponseEntity<String> handleUnhandledException(RuntimeException exception) {
-		return ResponseEntity.badRequest().body(exception.getMessage());
-	}
+    @ExceptionHandler
+    public ResponseEntity<String> handleUnhandledException(RuntimeException exception) {
+        return ResponseEntity.badRequest().body(exception.getMessage());
+    }
 }
