+++ src/main/java/woowacourse/auth/ui/CustomerController.java
@@ -1,9 +1,7 @@
 package woowacourse.auth.ui;
 
 import java.net.URI;
-
 import javax.validation.Valid;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,52 +11,53 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
-import lombok.RequiredArgsConstructor;
 import woowacourse.auth.application.CustomerService;
 import woowacourse.auth.domain.Customer;
-import woowacourse.auth.dto.CustomerRequest;
-import woowacourse.auth.dto.CustomerResponse;
-import woowacourse.auth.dto.CustomerUpdateRequest;
-import woowacourse.auth.dto.CustomerUpdateResponse;
+import woowacourse.auth.dto.customer.CustomerRequest;
+import woowacourse.auth.dto.customer.CustomerResponse;
+import woowacourse.auth.dto.customer.CustomerUpdateRequest;
+import woowacourse.auth.dto.customer.CustomerUpdateResponse;
 import woowacourse.auth.support.Login;
 
 @RestController
 @RequestMapping("/customers")
-@RequiredArgsConstructor
 public class CustomerController {
 
-	private final CustomerService customerService;
+    private final CustomerService customerService;
 
-	@PostMapping
-	public ResponseEntity<CustomerResponse> signUp(@RequestBody @Valid CustomerRequest customerRequest) {
-		Customer customer = customerService.signUp(customerRequest);
-		URI uri = createUri(customer.getId());
-		return ResponseEntity.created(uri).body(new CustomerResponse(customer));
-	}
+    public CustomerController(CustomerService customerService) {
+        this.customerService = customerService;
+    }
 
-	private URI createUri(Long id) {
-		return ServletUriComponentsBuilder
-			.fromCurrentRequest()
-			.path("/" + id)
-			.build().toUri();
-	}
+    @PostMapping
+    public ResponseEntity<CustomerResponse> signUp(@RequestBody @Valid CustomerRequest customerRequest) {
+        Customer customer = customerService.signUp(customerRequest);
+        URI uri = createUri(customer.getId());
+        return ResponseEntity.created(uri).body(new CustomerResponse(customer));
+    }
 
-	@DeleteMapping
-	public ResponseEntity<Void> signOut(@Login Customer customer) {
-		customerService.delete(customer);
-		return ResponseEntity.noContent().build();
-	}
+    @DeleteMapping
+    public ResponseEntity<Void> signOut(@Login Customer customer) {
+        customerService.delete(customer);
+        return ResponseEntity.noContent().build();
+    }
 
-	@PatchMapping
-	public ResponseEntity<CustomerUpdateResponse> update(@Login Customer customer,
-		@RequestBody @Valid CustomerUpdateRequest request) {
-		Customer updatedCustomer = customerService.update(customer, request);
-		return ResponseEntity.ok(new CustomerUpdateResponse(updatedCustomer.getNickname()));
-	}
+    @PatchMapping
+    public ResponseEntity<CustomerUpdateResponse> update(@Login Customer customer,
+                                                         @RequestBody @Valid CustomerUpdateRequest request) {
+        Customer updatedCustomer = customerService.update(customer, request);
+        return ResponseEntity.ok(new CustomerUpdateResponse(updatedCustomer.getNickname()));
+    }
 
-	@GetMapping
-	public ResponseEntity<CustomerResponse> find(@Login Customer customer) {
-		return ResponseEntity.ok(new CustomerResponse(customer));
-	}
+    @GetMapping
+    public ResponseEntity<CustomerResponse> find(@Login Customer customer) {
+        return ResponseEntity.ok(new CustomerResponse(customer));
+    }
+
+    private URI createUri(Long id) {
+        return ServletUriComponentsBuilder
+                .fromCurrentRequest()
+                .path("/" + id)
+                .build().toUri();
+    }
 }
