+++ src/main/java/woowacourse/auth/ui/config/AuthenticationPrincipalArgumentResolver.java
@@ -1,35 +1,37 @@
 package woowacourse.auth.ui.config;
 
 import javax.servlet.http.HttpServletRequest;
-
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-
-import lombok.RequiredArgsConstructor;
-import woowacourse.auth.domain.Customer;
 import woowacourse.auth.application.CustomerService;
+import woowacourse.auth.domain.Customer;
 import woowacourse.auth.support.AuthorizationExtractor;
 import woowacourse.auth.support.JwtTokenProvider;
 import woowacourse.auth.support.Login;
 
-@RequiredArgsConstructor
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
 
     private final JwtTokenProvider jwtTokenProvider;
     private final CustomerService customerService;
 
+    public AuthenticationPrincipalArgumentResolver(JwtTokenProvider jwtTokenProvider,
+                                                   CustomerService customerService) {
+        this.jwtTokenProvider = jwtTokenProvider;
+        this.customerService = customerService;
+    }
+
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
         return parameter.hasParameterAnnotation(Login.class)
-            && parameter.getParameterType().isAssignableFrom(Customer.class);
+                && parameter.getParameterType().isAssignableFrom(Customer.class);
     }
 
     @Override
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
-        NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
+                                  NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
         String email = jwtTokenProvider.getPayload(AuthorizationExtractor.extract(request));
         return customerService.findByEmail(email);
