+++ src/main/java/woowacourse/auth/ui/config/AuthenticationPrincipalConfig.java
@@ -1,28 +1,29 @@
 package woowacourse.auth.ui.config;
 
-import lombok.RequiredArgsConstructor;
-import woowacourse.auth.application.CustomerService;
-import woowacourse.auth.support.JwtTokenProvider;
-
+import java.util.List;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
+import woowacourse.auth.application.CustomerService;
+import woowacourse.auth.support.JwtTokenProvider;
 
 @Configuration
-@RequiredArgsConstructor
 public class AuthenticationPrincipalConfig implements WebMvcConfigurer {
 
     private final JwtTokenProvider jwtTokenProvider;
     private final CustomerService customerService;
 
+    public AuthenticationPrincipalConfig(JwtTokenProvider jwtTokenProvider, CustomerService customerService) {
+        this.jwtTokenProvider = jwtTokenProvider;
+        this.customerService = customerService;
+    }
+
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(loginInterceptor())
-            .excludePathPatterns("/auth/login");
+                .excludePathPatterns("/auth/login");
     }
 
     @Override
