+++ src/main/java/woowacourse/auth/ui/config/LoginInterceptor.java
@@ -2,41 +2,41 @@ package woowacourse.auth.ui.config;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
 import org.springframework.http.HttpStatus;
 import org.springframework.web.servlet.HandlerInterceptor;
-
-import lombok.RequiredArgsConstructor;
 import woowacourse.auth.support.AuthorizationExtractor;
 import woowacourse.auth.support.JwtTokenProvider;
 
-@RequiredArgsConstructor
 public class LoginInterceptor implements HandlerInterceptor {
 
-	private final JwtTokenProvider tokenProvider;
+    private final JwtTokenProvider tokenProvider;
 
-	@Override
-	public boolean preHandle(
-		HttpServletRequest request, HttpServletResponse response, Object handler) {
+    public LoginInterceptor(JwtTokenProvider tokenProvider) {
+        this.tokenProvider = tokenProvider;
+    }
 
-		if (isSignUpRequest(request)) {
-			return true;
-		}
+    @Override
+    public boolean preHandle(
+            HttpServletRequest request, HttpServletResponse response, Object handler) {
 
-		String token = AuthorizationExtractor.extract(request);
-		if (isUnauthorizedRequest(token)) {
-			response.setStatus(HttpStatus.UNAUTHORIZED.value());
-			return false;
-		}
-		return true;
-	}
+        if (isSignUpRequest(request)) {
+            return true;
+        }
 
-	private boolean isSignUpRequest(HttpServletRequest request) {
-		return request.getRequestURI().contains("/customer")
-			&& request.getMethod().equalsIgnoreCase("post");
-	}
+        String token = AuthorizationExtractor.extract(request);
+        if (isUnauthorizedRequest(token)) {
+            response.setStatus(HttpStatus.UNAUTHORIZED.value());
+            return false;
+        }
+        return true;
+    }
 
-	private boolean isUnauthorizedRequest(String nickname) {
-		return nickname == null || !tokenProvider.validateToken(nickname);
-	}
+    private boolean isSignUpRequest(HttpServletRequest request) {
+        return request.getRequestURI().contains("/customer")
+                && request.getMethod().equalsIgnoreCase("post");
+    }
+
+    private boolean isUnauthorizedRequest(String nickname) {
+        return nickname == null || !tokenProvider.validateToken(nickname);
+    }
 }
