+++ src/main/resources/application.properties
@@ -3,4 +3,4 @@ spring.datasource.url=jdbc:h2:~/test;MODE=MYSQL;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EX
 spring.datasource.username=sa
 spring.h2.console.enabled=true
 security.jwt.token.secret-key=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIiLCJuYW1lIjoiSm9obiBEb2UiLCJpYXQiOjE1MTYyMzkwMjJ9.ih1aovtQShabQ7l0cINw4k1fagApg3qLWiB8Kt59Lno
-security.jwt.token.expire-length= 3600000
+security.jwt.token.expire-length=3600000
